backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.2.3
gunicorn==21.2.0
mysqlclient==2.2.4
Flask-Cors==4.0.0
